# src/Common.Rest.Services/Address/Common.Rest.Address.Api/openapi.yaml
openapi: 3.0.0
info:
  title: Address API
  description: REST API for managing address records with CRUD operations, advanced search, and geographic data
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Apache 2.0

servers:
  - url: https://localhost:5011/api/v1
    description: Development server
  - url: http://localhost:5010/api/v1
    description: Development server (HTTP)

tags:
  - name: Address
    description: Address CRUD operations
  - name: Search
    description: Search and filter address records
  - name: Lifecycle
    description: Address lifecycle management (delete, restore)

paths:
  /Address:
    get:
      tags:
        - Address
      summary: Get all addresses with pagination
      description: Retrieve a paginated list of all addresses
      operationId: getAllAddresses
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated list of addresses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedApiResponse'
              examples:
                success:
                  value:
                    success: true
                    message: "Request completed successfully"
                    data:
                      items:
                        - id: 1
                          addressDetails:
                            uprn: "123456789012"
                            address:
                              organisation: "Test Organisation Ltd"
                              department: "Planning Department"
                              pao:
                                text: "10 Downing Street"
                                startNumber: 10
                                endNumber: 12
                              streetDescriptor:
                                streetDescription: "High Street"
                                locality: "Central London"
                                postTown: "Westminster"
                              postcode: "SW1A 1AA"
                            geography:
                              easting: 530000
                              northing: 180000
                      pageNumber: 1
                      pageSize: 10
                      totalRecords: 50
                      totalPages: 5
                    correlationId: "correlation-trace-123456"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Address
      summary: Create a new address
      description: Create a new address with geographic coordinates and detailed address information
      operationId: createAddress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateAddress'
            examples:
              valid:
                value:
                  address:
                    organisation: "Test Organisation Ltd"
                    department: "Planning Department"
                    sao:
                      text: "Flat 1"
                      startNumber: 1
                      startSuffix: "A"
                    pao:
                      text: "10 Downing Street"
                      startNumber: 10
                      endNumber: 12
                    streetDescriptor:
                      streetDescription: "High Street"
                      locality: "Central London"
                      dependentLocality: "Westminster"
                      townName: "London"
                      postTown: "Westminster"
                      administrativeArea: "Greater London"
                    postcode: "SW1A 1AA"
                  geography:
                    easting: 530000
                    northing: 180000
              minimal:
                value:
                  address:
                    pao:
                      text: "5 High Street"
                    streetDescriptor:
                      streetDescription: "Main Street"
                  geography:
                    easting: 500000
                    northing: 150000
      responses:
        '201':
          description: Address created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                created:
                  value:
                    success: true
                    message: "Address created successfully"
                    data:
                      id: 1
                      addressDetails:
                        uprn: "123456789012"
                        address:
                          organisation: "Test Organisation Ltd"
                          pao:
                            text: "10 Downing Street"
                            startNumber: 10
                            endNumber: 12
                          streetDescriptor:
                            streetDescription: "High Street"
                          postcode: "SW1A 1AA"
                        geography:
                          easting: 530000
                          northing: 180000
                    correlationId: "correlation-trace-123456"
        '400':
          description: Invalid request data (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                missingRequired:
                  value:
                    success: false
                    message: "Validation failed"
                    errors:
                      - field: "address"
                        message: "Address information is required"
                      - field: "geography"
                        message: "Geographic coordinates are required"
                invalidPostcode:
                  value:
                    success: false
                    message: "Validation failed"
                    errors:
                      - field: "address.postcode"
                        message: "Invalid postcode format"
                invalidEasting:
                  value:
                    success: false
                    message: "Validation failed"
                    errors:
                      - field: "geography.easting"
                        message: "Easting must be between 0 and 2147483647"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /Address/{id}:
    get:
      tags:
        - Address
      summary: Get address by ID
      description: Retrieve a single address by its unique identifier
      operationId: getAddressById
      parameters:
        - name: id
          in: path
          required: true
          description: Address unique identifier (GUID)
          schema:
            type: string
            format: uuid
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Invalid GUID format provided"
                correlationId: "correlation-trace-123456"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Address
      summary: Update an address
      description: Update an existing address (full or partial update)
      operationId: updateAddress
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Address unique identifier (GUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateAddress'
            examples:
              fullUpdate:
                value:
                  address:
                    organisation: "Updated Organisation Ltd"
                    pao:
                      text: "20 Downing Street"
                      startNumber: 20
                    streetDescriptor:
                      streetDescription: "Updated Street"
                      postTown: "Updated Town"
                    postcode: "SW1A 2AA"
                  geography:
                    easting: 531000
                    northing: 181000
              partialUpdate:
                value:
                  address: null
                  geography:
                    easting: 532000
                    northing: 182000
      responses:
        '200':
          description: Address updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Lifecycle
      summary: Soft delete an address
      description: Soft delete (mark as deleted) an address record
      operationId: deleteAddress
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Address unique identifier (GUID)
          schema:
            type: string
            format: uuid
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Address soft deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /Address/{id}/permanent:
    delete:
      tags:
        - Lifecycle
      summary: Permanently delete an address
      description: Permanently delete an address record from the database
      operationId: permanentlyDeleteAddress
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Address unique identifier (GUID)
          schema:
            type: string
            format: uuid
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Address permanently deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /Address/{id}/restore:
    post:
      tags:
        - Lifecycle
      summary: Restore a deleted address
      description: Restore a previously soft-deleted address
      operationId: restoreAddress
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Address unique identifier (GUID)
          schema:
            type: string
            format: uuid
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Address restored successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /Address/search/uprn:
    get:
      tags:
        - Search
      summary: Search address by UPRN
      description: Find an address by its Unique Property Reference Number (UPRN)
      operationId: searchByUprn
      parameters:
        - name: uprn
          in: query
          required: true
          description: Unique Property Reference Number
          schema:
            type: string
            minLength: 1
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: UPRN is required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Address with specified UPRN not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Address not found with UPRN: 999999999999"
                correlationId: "correlation-trace-123456"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /Address/search/advanced:
    get:
      tags:
        - Search
      summary: Advanced search with multiple criteria
      description: Search addresses using multiple optional criteria (postcode, postTown, organisation, thoroughfare, locality). Results are filtered by AND logic.
      operationId: advancedSearch
      parameters:
        - name: postcode
          in: query
          description: UK postcode to search
          schema:
            type: string
          example: "SW1A 1AA"
        - name: postTown
          in: query
          description: Post town to search (case-insensitive)
          schema:
            type: string
          example: "Westminster"
        - name: organisation
          in: query
          description: Organisation name to search (case-insensitive)
          schema:
            type: string
          example: "Test Organisation"
        - name: thoroughfare
          in: query
          description: Street description to search (case-insensitive)
          schema:
            type: string
          example: "High Street"
        - name: locality
          in: query
          description: Locality to search (case-insensitive)
          schema:
            type: string
          example: "Central London"
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Search results returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedApiResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /Address/count:
    get:
      tags:
        - Search
      summary: Get total address count
      description: Get the total number of non-deleted addresses in the system
      operationId: getAddressCount
      parameters:
        - name: X-Correlation-ID
          in: header
          description: Correlation ID for tracing
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Total count retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: integer
                    description: Total count of addresses
                  correlationId:
                    type: string
                    format: uuid
              example:
                success: true
                message: "Request completed successfully"
                data: 42
                correlationId: "correlation-trace-123456"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication. Replace 'your-bearer-token-here' with actual JWT token.

  schemas:
    CreateUpdateAddress:
      type: object
      required:
        - address
        - geography
      properties:
        address:
          $ref: '#/components/schemas/AddressInfo'
          nullable: true
        geography:
          $ref: '#/components/schemas/Geography'
          nullable: true

    AddressInfo:
      type: object
      required:
        - pao
        - streetDescriptor
      properties:
        organisation:
          type: string
          maxLength: 120
          nullable: true
          description: Organisation name
        department:
          type: string
          maxLength: 120
          nullable: true
          description: Department within organisation
        sao:
          $ref: '#/components/schemas/AddressableObject'
          nullable: true
          description: Secondary Addressable Object (e.g., Flat number)
        pao:
          $ref: '#/components/schemas/AddressableObject'
          description: Primary Addressable Object (e.g., House number)
        streetDescriptor:
          $ref: '#/components/schemas/StreetDescriptor'
        postcode:
          type: string
          pattern: '^(GIR 0AA|([A-PR-UWYZ][0-9][0-9]?|[A-PR-UWYZ][A-HK-Y][0-9][0-9]?|[A-PR-UWYZ][0-9][A-HJKSTUW]|[A-PR-UWYZ][A-HK-Y][0-9][ABEHMNPRVWXY]) [0-9][ABD-HJLNP-UW-Z]{2})$'
          nullable: true
          description: UK postcode (or special format GIR 0AA)
          example: "SW1A 1AA"

    AddressableObject:
      type: object
      properties:
        text:
          type: string
          maxLength: 120
          description: Full text representation (e.g., "10 Downing Street")
        startNumber:
          type: integer
          minimum: 1
          maximum: 99999
          nullable: true
          description: Starting house/building number
        startSuffix:
          type: string
          pattern: '^[A-Za-z]$'
          nullable: true
          description: Starting number suffix (single letter)
        endNumber:
          type: integer
          minimum: 1
          maximum: 99999
          nullable: true
          description: Ending house/building number for ranges
        endSuffix:
          type: string
          pattern: '^[A-Za-z]$'
          nullable: true
          description: Ending number suffix (single letter)

    StreetDescriptor:
      type: object
      required:
        - streetDescription
      properties:
        streetDescription:
          type: string
          maxLength: 120
          description: Name of the street/thoroughfare
          example: "High Street"
        locality:
          type: string
          maxLength: 120
          nullable: true
          description: Locality name
        dependentLocality:
          type: string
          maxLength: 120
          nullable: true
          description: Dependent locality
        doubleDependentLocality:
          type: string
          maxLength: 120
          nullable: true
          description: Double dependent locality
        townName:
          type: string
          maxLength: 120
          nullable: true
          description: Town name
        postTown:
          type: string
          maxLength: 120
          nullable: true
          description: Post town
        administrativeArea:
          type: string
          maxLength: 120
          nullable: true
          description: Administrative area

    Geography:
      type: object
      required:
        - easting
        - northing
      properties:
        easting:
          type: integer
          minimum: 0
          maximum: 2147483647
          description: Ordnance Survey Easting coordinate
          example: 530000
        northing:
          type: integer
          minimum: 0
          maximum: 2147483647
          description: Ordnance Survey Northing coordinate
          example: 180000

    AddressDocument:
      type: object
      properties:
        id:
          type: integer
          description: Database-generated identifier
        addressDetails:
          type: object
          properties:
            uprn:
              type: string
              description: Unique Property Reference Number
            address:
              $ref: '#/components/schemas/AddressInfo'
            geography:
              $ref: '#/components/schemas/Geography'

    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        message:
          type: string
          description: Response message
        data:
          oneOf:
            - $ref: '#/components/schemas/AddressDocument'
            - type: array
          nullable: true
          description: Response data (address or array of addresses)
        correlationId:
          type: string
          format: uuid
          description: Correlation ID for request tracing

    PagedApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AddressDocument'
            pageNumber:
              type: integer
              description: Current page number
            pageSize:
              type: integer
              description: Number of items per page
            totalRecords:
              type: integer
              description: Total number of records available
            totalPages:
              type: integer
              description: Total number of pages
        correlationId:
          type: string
          format: uuid

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        correlationId:
          type: string
          format: uuid

    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        correlationId:
          type: string
          format: uuid

  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Address not found"
            correlationId: "correlation-trace-123456"

    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Unauthorized access"
            correlationId: "correlation-trace-123456"

    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'